requirements.txt
gspread==5.5.0
python-dotenv==0.21.0
selenium==3.141.0
beautifulsoup4==4.11.1
geckodriver-autoinstaller==0.1.0
python-dateutil==2.8.2
python-telegram-bot==13.12
